Move code from advancedsettings.h to its own cpp file
